# src/swagger.json
{
	"components": {
...
		"schemas": {
			"FieldErrors": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"properties": {
						"value": {},
						"message": {
							"type": "string"
						}
					},
					"required": [
						"message"
					],
					"type": "object"
				}
			},
...
			"VolunteerError": {
				"properties": {
					"name": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ErrorName"
							},
							{
								"type": "string",
								"enum": [
									"UNSPECIFIED_ERROR"
								]
							}
						]
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"details": {}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PaginationResult_NotebookWithPlaceAndVolunteer-Array_": {
				"properties": {
					"totalCount": {
						"type": "number",
						"format": "double"
					},
					"pageInfo": {
						"properties": {
							"hasPreviousPage": {
								"type": "boolean"
							},
							"hasNextPage": {
								"type": "boolean"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"hasPreviousPage",
							"hasNextPage",
							"page"
						],
						"type": "object"
					},
					"nodes": {
						"items": {
							"$ref": "#/components/schemas/NotebookWithPlaceAndVolunteer"
						},
						"type": "array"
					}
				},
				"required": [
					"totalCount",
					"pageInfo",
					"nodes"
				],
				"type": "object"
			},
			"NotebookError": {
				"properties": {
					"name": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ErrorName"
							},
							{
								"type": "string",
								"enum": [
									"UNSPECIFIED_ERROR"
								]
							}
						]
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"details": {}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EvaluateNotebookEntity": {
				"$ref": "#/components/schemas/Omit_%2F%2Feslint-disable-next-lineprettier%2FprettierNotebookEntity.-or-idcad-or-studentName-or-studentRegistration-or-idpep-or-reservationDate-or-evaluatedDate-or-notebookDirectory_"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "palavrasdepaz",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/notebooks/evaluation-list": {
			"get": {
				"operationId": "GetNotebooksEvaluation",
				"responses": {
					"200": {
						"description": "Successfully fetched the notebooks",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginationResult_NotebookWithPlaceAndVolunteer-Array_"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "List all notebooks.\n Pagination\n Page: ?page=number& (page number)\n Limit: ?limit=number& (data quantity - max=30)\n Filter: ?classes=399,404,405& (filter by class ids - separator \",\" (comma))",
				"tags": [
					"Notebook"
				],
				"security": [
					{
						"jwt": [
							"readPermission"
						]
					}
				],
				"parameters": []
			}
		},
		"/notebooks/evaluation-list/download": {
			"get": {
				"operationId": "GetNotebooksEvaluationDownload",
				"responses": {
					"200": {
						"description": "Successfully fetched the notebooks",
						"content": {
							"application/json": {
								"schema": {
									"type": "string",
									"format": "byte"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Download  all notebooks.\n Filter: ?classes=399,404,405& (filter by class ids - separator \",\" (comma))",
				"tags": [
					"Notebook"
				],
				"security": [
					{
						"jwt": [
							"readPermission"
						]
					}
				],
				"parameters": []
			}
		},
		"/notebooks/count/{idvol}": {
			"get": {
				"operationId": "CountEvaluatedNotebooksByIdVol",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"count": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"count"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Get total count of notebooks by a volunteer.",
				"tags": [
					"Notebook"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "idvol",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/notebooks/download/{notebookId}": {
			"get": {
				"operationId": "DownloadNotebookFromId",
				"responses": {
					"200": {
						"description": "Successfully downloaded the files",
						"content": {
							"application/json": {
								"schema": {
									"type": "string",
									"format": "byte"
								}
							}
						}
					},
					"404": {
						"description": "Essay not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotebookError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "NOTEBOOK_NOT_FOUND_ERROR",
											"message": "Notebook with id {some Id} not found"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Download the notebook with specified id as a pdf file\nThe return of this route is a stream (content type: application/octet-stream)\nand the response header include the file as an attachement\n\n(The volunteer must have readPermission, which is checked using JWT)",
				"tags": [
					"Notebook"
				],
				"security": [
					{
						"jwt": [
							"readPermission"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "notebookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/notebooks/available/{idvol}": {
			"get": {
				"operationId": "GetAvailableNotebooksForEvalForIdVol",
				"responses": {
					"200": {
						"description": "Successfully fetched the notebooks",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AvailableNotebookRowEntity"
									},
									"type": "array"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Get available notebooks for evaluation for the volunteer,\nthose notebooks includes the ones which does not have a reservation date\nor the reservations of the volunteer.\n\n(The volunteer must have readPermission, which is checked using JWT)",
				"tags": [
					"Notebook"
				],
				"security": [
					{
						"jwt": [
							"readPermission"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "idvol",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/notebooks/evaluation/{notebookId}": {
			"put": {
				"operationId": "SaveNotebookEvaluation",
				"responses": {
					"200": {
						"description": "Successfully Evaluated notebook",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotebookEntity"
								}
							}
						}
					},
					"400": {
						"description": "Notebook already evaluated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotebookError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "NOTEBOOK_ALREADY_RESERVED_ERROR",
											"message": "Notebook with id {some notebook id} not found"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Notebook not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotebookError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "NOTEBOOK_NOT_FOUND_ERROR",
											"message": "Notebook with id {some notebook id} not found"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Evaluate a notebook\n\n(The volunteer must have readPermission, which is checked using JWT)",
				"tags": [
					"Notebook"
				],
				"security": [
					{
						"jwt": [
							"readPermission"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "notebookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EvaluateNotebookEntity"
							}
						}
					}
				}
			}
		},
		"/notebooks/reservation": {
			"put": {
				"operationId": "ReserveNotebookForVolunteer",
				"responses": {
					"200": {
						"description": "Successfully reserved notebook for volunteer",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AvailableNotebookRowEntity"
								}
							}
						}
					},
					"400": {
						"description": "Notebook already reserved or evaluated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotebookError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "NOTEBOOK_ALREADY_RESERVED_ERROR",
											"message": "Notebook already reserved or already evaluated"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Notebook not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotebookError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "NOTEBOOK_NOT_FOUND_ERROR",
											"message": "Notebook with id {some notebook id} not found"
										}
									}
								}
							}
						}
					},
					"412": {
						"description": "Volunteer not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VolunteerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "VOLUNTEER_NOT_FOUND",
											"message": "Volunteer with id {some volunteer id} not found"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Reserve notebook for the volunteer. If the notebook is already reserve or evaluated\nstatus 400 is returned.\n\n(The volunteer must have readPermission, which is checked using JWT)",
				"tags": [
					"Notebook"
				],
				"security": [
					{
						"jwt": [
							"readPermission"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReserveNotebookDataEntity"
							}
						}
					}
				}
			}
		},
		"/notebooks/revert-reservation/{notebookId}": {
			"put": {
				"operationId": "RevertReserveNotebookForVolunteer",
				"responses": {
					"200": {
						"description": "Successfully revert reserved notebook for volunteer",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AvailableNotebookRowEntity"
								}
							}
						}
					},
					"404": {
						"description": "Notebook not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotebookError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "NOTEBOOK_NOT_FOUND_ERROR",
											"message": "Notebook with id {some notebook id} not found"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Revert reserve notebook for the volunteer.",
				"tags": [
					"Notebook"
				],
				"security": [
					{
						"jwt": [
							"readPermission"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "notebookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}